Direction Indicator and Follow Solver Graduation (#9020)
* updated solver with dynamic scale, made adjustements to the demo scene

* updated unit tests to account for scaling indicator size

* updating the Directional Solver documentation language

* refactored solvers out of their experimental folders and into mainline, added Direction Indicator documentation

* added documentation for the follow solver

* update release notes

* fixing broken links

* fixing broken bookmark
