--- Assets/MRTK/Examples/Experimental/Solvers/DirectionalIndicatorExample.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4c78d9cf1eaba1d42b902e09f6eabf6e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
