--- Assets/MRTK/Examples/Experimental/Solvers/FollowSolverExample.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f2d4fc481a94ae44684441bd328081d6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
