--- Assets/MRTK/SDK/Editor/Inspectors/Experimental/FollowSolver.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c93333473212a8e4ebb8ece7960e553c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
