--- Assets/MRTK/SDK/Editor/Inspectors/Experimental/FollowSolver/FollowInspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 836b0f72b07e57a4ca58c2a54f7e2e6a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
