--- Assets/MRTK/SDK/Experimental/Features/Utilities/DirectionalIndicator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c45719620cce37b49bae491053c48bed
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
