--- Assets/MRTK/SDK/Experimental/FollowSolver.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e59efff6263f6de41940081cb78d9d25
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
