+++ Assets/MRTK/SDK/Features/Utilities/Solvers/Follow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4479a0bd44d822241a4c5f4890c481c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
