+++ Assets/MRTK/SDK/Features/Utilities/Solvers/Solver.cs
@@ -68,8 +68,8 @@ namespace Microsoft.MixedReality.Toolkit.Utilities.Solvers
         }
 
         [SerializeField]
-        [Tooltip("If true, the Solver will respect the object's original scale values")]
-        private bool maintainScale = true;
+        [Tooltip("If true, the Solver will respect the object's original scale values on initialization")]
+        private bool maintainScaleOnInitialization = true;
 
         [SerializeField]
         [Tooltip("If true, updates are smoothed to the target. Otherwise, they are snapped to the target")]
@@ -206,7 +206,7 @@ namespace Microsoft.MixedReality.Toolkit.Utilities.Solvers
                 updateLinkedTransform = false;
             }
 
-            GoalScale = maintainScale ? transform.localScale : Vector3.one;
+            GoalScale = maintainScaleOnInitialization ? transform.localScale : Vector3.one;
         }
 
         /// <summary>
